+++ packages/EasyApiPlatform/tests/Stubs/app/src/ApiResource/RelatedDummy.php
@@ -13,7 +13,7 @@ use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
-use EonX\EasyApiPlatform\Filter\AdvancedSearchFilter;
+use EonX\EasyApiPlatform\Filters\AdvancedSearchFilter;
 use Symfony\Component\Serializer\Annotation\Groups;
 use Symfony\Component\Validator\Constraints as Assert;
